Get rid of SyntaxErrors in PerformanceObserver
This CL changes the SyntaxErrors to TypeErrors when 'type' and
'entryTypes' are both present or not present. It also adds a UseCounter
for when 'entryTypes' is present and 'buffered', the only other member
in the dictionary, is also present. Currently this causes a console
warning whereas it should instead also throw a TypeError, so once this
mistake is proved to be very infrequent, we can proceed to change it to
TypeError. See related PR:

https://github.com/w3c/performance-timeline/pull/164

Change-Id: Ic9399bf54895e3723b0ca2e513161a7e0638b681